// src/types.ts
export enum Platform {
  YouTube = "YouTube",
  TikTok = "TikTok",
  Instagram = "Instagram",
  Twitter = "Twitter X",
  LinkedIn = "LinkedIn",
  Facebook = "Facebook",
}